# Conflicts:
#	resources/views/admin/detailTalent.blade.php

+++ app/Http/Controllers/TalentNewController.php
@@ -6,9 +6,11 @@ use Illuminate\Http\Request;
 
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Mail;
+use Maatwebsite\Excel\Concerns\WithHeadings;
 
 use App\Mail\UpscaleEmail;
 use App\Models\Talent;
+use App\Models\Skill;
 use App\Models\Talent_log;
 use App\User;
 
@@ -20,6 +22,8 @@ use App\Http\Controllers\Controller;
 
 use Illuminate\Database\Eloquent\Builder;
 
+use Illuminate\Support\Facades\Crypt;
+
 class TalentNewController extends Controller
 {
 
@@ -29,9 +33,23 @@ class TalentNewController extends Controller
         return view('admin.talent.home');
     }
 
-    function mail($id)
+    function mail($talent_id)
     {
-        return view("admin.talent.mail");
+        $talent = Talent::findOrFail($talent_id);
+        return view("admin.talent.mail",compact('talent'));
+    }
+
+    function createTypeEmail($id)
+    {
+        $talent = Talent::find($id);
+        return view("admin.talent.createTypeEmail",compact('talent'));
+    }
+
+    function mailBackup($id)
+    {
+        $talent = Talent::find($id); 
+        return view("admin.talent.mail-backup",compact('talent'));
+
     }
 
     public function mailSend(Request $request)
@@ -89,15 +107,30 @@ class TalentNewController extends Controller
 
     public function paginate_data(Request $request)
     {
+        //DB::enableQueryLog();
+
         // if ($request->ajax()) {
 
             //SELECT BUILDER START
-            $default_query = "*, users.email as member_email, users.created_at as member_date";
+            $default_query = "*,users.id as user_id, users.email as member_email, users.created_at as member_date";
+            
+            if ( $request->jumlah_apply_jobs)
+            {
+                $default_query .=", COUNT(jobs_apply_id) as jumlah_apply_jobs"; 
+            }
+            else
+            {
+                $default_query .=", '-' as jumlah_apply_jobs"; 
+            }
             $data = Talent::select(DB::raw($default_query));
             //SELECT BUILDER END 
 
             //JOIN BUILDER START
             $data->join("users","talent.user_id","=","users.id","LEFT");
+            if ( $request->jumlah_apply_jobs || $request->apply == 'yes')
+            {
+                $data->join('jobs_apply','jobs_apply.jobs_apply_talent_id','=','talent.talent_id',"LEFT");
+            }
             //JOIN BULDER END 
 
             if ( $request->talent_name ) {$data->where("talent_name","LIKE","%".$request->talent_name."%"); }
@@ -107,6 +140,7 @@ class TalentNewController extends Controller
             if ( $request->talent_onsite_jogja ) {$data->where("talent_onsite_jogja",$request->talent_onsite_jogja); }
             if ( $request->talent_onsite_jakarta ) {$data->where("talent_onsite_jakarta",$request->talent_onsite_jakarta); }
             if ( $request->talent_isa ) {$data->where("talent_isa",$request->talent_isa); }
+            if ( $request->talent_focus ) {$data->where("talent_focus","LIKE","%".$request->talent_focus."%"); }
 
             if ( $request->status_member == "member" )
             {
@@ -117,6 +151,12 @@ class TalentNewController extends Controller
             {
                 $data->where("users.email","=",null);
             }
+
+            if ( $request->apply == 'yes' )
+            {
+                $data->having(DB::raw('COUNT(jobs_apply_id)'),'>',0);
+            }
+
             if ( $request->order != '' )
             {
                 $ar = explode(",",$request->order);
@@ -129,14 +169,16 @@ class TalentNewController extends Controller
             {
                 $data->orderBy("talent_id","DESC");
             }
-
+            $data->groupBy("talent_id");
             $data = $data->paginate(10);
+
+            //$query = DB::getQueryLog(); dd($query);
 
             return view('admin.talent.table',compact('data'))->render();
         // }
     }
 
-    public function export_excel()
+    public function export_excel ()
     {
         return Excel::download(new TalentExport, 'talent.xlsx');
     }
@@ -155,8 +197,11 @@ class TalentNewController extends Controller
             $talent = Talent::find($row);
             
             //menghapus semua data di table user yg berelasi
-            $user = User::find($talent->user_id); 
-            $user->delete() ; 
+            if ( $user = User::find($talent->user_id) ) 
+            {
+
+                $user->delete() ; 
+            } 
 
             //delete 
             Talent::where('talent_id', $row)->delete();
@@ -168,60 +213,61 @@ class TalentNewController extends Controller
         return view('admin.talent.insert');
     }
 
-public function insertData(Request $request){
+    public function insertData(Request $request){
 
 
 
         $validation = $request->validate([
             'nama'=>'required|string|max:150',
             'email'=>'required|string|email|max:100|unique:users',
-            'gender'=>'required',
-            'alamat'=>'required',
-            'phone'=>'required|string|max:30',
-            'birthdate'=>'required',
-            'birthplace'=>'required',
-            'martialstatus'=>'required',
-            'currentaddress'=>'required|string',
-            'condition'=>'required',
-            'skill'=>'required',
-            'salary'=>'required|string',
-            'focus'=>'required|string',
-            'startcareer'=>'required|string',
-            'level'=>'required',
-            'latestsalary'=>'required|string',
-            'preflocation'=>'required|string',
-            'status'=>'required',
-            'onsite'=>'required',
-            'remote'=>'required',
-            'available'=>'required',
-            'apply'=>'required',
-            'international'=>'required',
-            'freelancehour'=>'required',
-            'projectmin'=>'required',
-            'projectmax'=>'required',
-            'konsulrate'=>'required',
-            'tutorrate'=>'required',
+            // 'gender'=>'required',
+            // 'alamat'=>'required',
+            // 'phone'=>'required|string|max:30',
+            // 'birthdate'=>'required',
+            // 'birthplace'=>'required',
+            // 'martialstatus'=>'required',
+            // 'currentaddress'=>'required|string',
+            // 'condition'=>'required',
+            // 'skill'=>'required',
+            // 'salary'=>'required|string',
+            // 'focus'=>'required|string',
+            // 'startcareer'=>'required|string',
+            // 'level'=>'required',
+            // 'latestsalary'=>'required|string',
+            // 'preflocation'=>'required|string',
+            // 'status'=>'required',
+            // 'onsite'=>'required',
+            // 'remote'=>'required',
+            // 'available'=>'required',
+            // 'apply'=>'required',
+            // 'international'=>'required',
+            // 'freelancehour'=>'required',
+            // 'projectmin'=>'required',
+            // 'projectmax'=>'required',
+            // 'konsulrate'=>'required',
+            // 'tutorrate'=>'required',
 
         ]);
 
+
+        
 
         DB::table('talent')->insert([
             'talent_name' => $request->nama,
             'talent_email' => $request->email,
             'talent_gender' => $request->gender,
             'talent_address' => $request->alamat,
-            'talent_phone' => $request->phone,
+            'talent_phone' => isset($request->phone) ? $request->phone : '',
             'talent_birth_date' => $request->birthdate,
             'talent_place_of_birth' => $request->birthplace,
             'talent_martial_status' => $request->martialstatus,
-            'talent_current_address' => $request->currentaddress,
+            'talent_current_address' => isset($request->currentaddress)?$request->currentaddress:'',
             'talent_condition' => $request->condition,
-            'talent_skill' => $request->skill,
             'talent_salary' => $request->salary,
             'talent_focus' => $request->focus,
             'talent_start_career' => $request->startcareer,
-            'talent_level' => $request->level,
-            'talent_latest_salary' => $request->latestsalary,
+            'talent_level' => isset($request->level)?$request->level:"undefined",
+            'talent_lastest_salary' => $request->latestsalary,
             'talent_prefered_location' => $request->preflocation,
             'talent_status' => $request->status,
             'talent_onsite_jakarta' => $request->onsite,
@@ -230,16 +276,40 @@ public function insertData(Request $request){
             'talent_apply' => $request->apply,
             'talent_international' => $request->international,
             'talent_location_id' => '12',
-            'talent_freelance_hour' => $request->freelancehour,
-            'talent_project_min' => $request->projectmin,
-            'talent_project_max' => $request->projectmax,
-            'talent_konsultasi_rate' => $request->konsulrate,
-            'talent_ngajar_rate' => $request->tutorrate,
+            'talent_freelance_hour' => isset($request->freelancehour)?$request->freelancehour:'',
+            'talent_project_min' => isset($request->projectmin)? $request->projectmin: '' ,
+            'talent_project_max' => isset($request->projectmax)? $request->projectmax: '' ,
+            'talent_konsultasi_rate' => isset($request->konsulrate)? $request->konsulrate: '' ,
+            'talent_ngajar_rate' => isset($request->tutorrate)? $request->tutorrate: '' ,
         ]);
 
+        $idTalent = DB::table('talent')->insertGetId([ 'talent_name' => $request->nama ]);
 
-        return redirect('/list');
 
+
+        if ( $request->skill != '' )
+        {
+            $InsertSkill = explode(",",$request->skill);
+            foreach ( $InsertSkill as $skill)
+            {
+                $idSkill[] = Skill::where("skill_name",$skill)->first()->skill_id;
+            }
+            foreach($idSkill as $insert)
+            {
+
+                DB::table('skill_talent')->insert([
+                'st_talent_id' => $idTalent,
+                'st_skill_id' => $insert,
+                'st_skill_verified' => 'NO',
+                ]);
+            
+            }
+
+        }
+
+        
+
+        return redirect('admin/talent/list/insert')->with('success', 'Data Talent Berhasil dimasukkan.');
        
     }
